Move grunt tasks to NPM
uglify, copy, webpack, karma
