+++ package.json
@@ -3,27 +3,31 @@
   "version": "2.0.8",
   "description": "TransfeWise Styleguide Components",
   "scripts": {
-    "pretest": "bower install",
-    "test": "grunt karma",
-    "prestart": "bower install",
+    "pretest": "",
+    "test": "karma start karma-conf.js --single-run",
+    "prestart": "npm install",
     "start": "grunt",
     "dev": "live-server --port=8181 --open=demo/index.html",
     "htmllint": "htmllint src/**/*.html; htmllint demo/**/*.html; exit 0",
     "eslint": "eslint src; exit 0",
+    "lint": "npm run htmllint && npm run eslint",
     "less": "lessc demo/**/*.less demo/css/examples.css",
-    "copy-jquery-js": "copyfiles -f node_modules/jquery/dist/jquery.min.js demo/lib/",
-    "copy-angular-js": "copyfiles -f node_modules/angular/angular.min.js demo/lib/",
-    "copy-bootstrap-js": "copyfiles -f node_modules/bootstrap/dist/js/bootstrap.min.js demo/lib/",
-    "copy-bootstrap-css": "copyfiles -f node_modules/bootstrap/dist/css/bootstrap.css demo/lib/",
-    "copy-flags-css": "copyfiles -f node_modules/currency-flags/dist/currency-flags.css demo/lib/",
-    "copy-demo-js": "npm run copy-jquery-js; npm run copy-angular-js; npm run copy-bootstrap-js;",
-    "copy-demo-css": "npm run copy-bootstrap-css; npm run copy-flags-css;",
-    "copy-lib": "copyfiles -f dist/js/styleguide-components.js demo/lib/",
-    "copy": "npm run copy-demo-css; npm run copy-demo-js;",
-    "uglify": "grunt uglify",
+    "copy-jq-js": "cp node_modules/jquery/dist/jquery.min.js demo/lib/",
+    "copy-ang-js": "cp node_modules/angular/angular.min.js demo/lib/",
+    "copy-bs-js": "cp node_modules/bootstrap/dist/js/bootstrap.min.js demo/lib/",
+    "copy-bs-css": "cp node_modules/bootstrap/dist/css/bootstrap.css demo/lib/",
+    "copy-flags-css": "cp node_modules/currency-flags/dist/currency-flags.css demo/lib/",
+    "copy-demo-fonts": "cp node_modules/bootstrap/fonts/*.* demo/fonts; cp node_modules/iconfont/fonts/*.* demo/fonts",
+    "copy-demo-js": "npm run copy-jq-js && npm run copy-ang-js && npm run copy-bs-js;",
+    "copy-demo-css": "npm run copy-bs-css && npm run copy-flags-css;",
+    "copy-demo": "npm run copy-demo-css && npm run copy-demo-js && npm run copy-demo-fonts;",
+    "copy-lib": "cp dist/js/styleguide-components.js demo/lib/",
+    "copy-ghpages": "cp -R demo/ gh-pages",
+    "uglify": "uglifyjs dist/js/styleguide-components.js -o dist/js/styleguide-components.min.js -c -m",
     "watch": "grunt watch",
-    "npm-start": "npm run htmllint; npm run eslint; npm run copy;",
-    "npm-build": "npm run htmllint; npm run eslint; npm run copy-lib;"
+    "npm-start": "webpack && npm run uglify && npm run less && npm run copy-demo;",
+    "npm-build": "webpack && npm run uglify && npm run less && npm run copy-lib && npm run copy-ghpages;",
+    "deploy": "git-directory-deploy --directory demo --branch gh-pages --message 'Update docs'"
   },
   "repository": {
     "type": "git",
@@ -51,12 +55,12 @@
     "babel-preset-es2015": "^6.24.1",
     "bootstrap": "git+https://github.com/transferwise/bootstrap.git",
     "bower": "*",
-    "copyfiles": "^1.2.0",
     "currency-flags": "git+https://github.com/transferwise/currency-flags.git",
     "eslint": "^4.6.0",
     "eslint-config-airbnb-base": "^12.0.0",
     "eslint-loader": "^1.9.0",
     "eslint-plugin-import": "^2.7.0",
+    "git-directory-deploy": "^1.5.1",
     "grunt": "^1.0.1",
     "grunt-bower-task": "^0.4.0",
     "grunt-contrib-copy": "^1.0.0",
@@ -78,6 +82,7 @@
     "jshint-loader": "^0.8.4",
     "karma": "^0.13.21",
     "karma-chrome-launcher": "^0.2.2",
+    "karma-cli": "^1.0.1",
     "karma-coverage": "^0.5.5",
     "karma-firefox-launcher": "^0.1.7",
     "karma-jasmine": "^0.3.7",
@@ -87,6 +92,7 @@
     "less": "^2.7.2",
     "live-server": "^1.1.0",
     "phantomjs-prebuilt": "^2.1.5",
+    "uglify-js": "^3.1.1",
     "webpack": "^2.5.1",
     "webpack-dev-server": "^2.4.5"
   }
